NavBar Created
Navigation Bar created. First followed an outdated React Router/MUI Components tutorial then a wild Tim appeared to help me change the components to a newer version of React. For now, I just made sure the functionality worked so there's no particular styling on it, but that will be added in the coming week.