custom carousel created + banner
Right now, the carousel animations just have hard-coded values for the animation, but eventually should have not hardcoded values according to the formula in customCarousel.css. Banner has also been cropped according to  figma specifications mostly.... but we can change that if needed. Will have to work tomorrow on state re-rendering of components