Changed CSS and added postcardPage
So I discovered the vw unit. It makes things a lot more scalable on a page and you can use pinch zooming to zoom in on any modern touchpad or tablet device.  I also then made that change for everything else in the navBar and the homepage. Oh and MUI can go. I don't use it with my things anymore