Some front-end enhancements
Client side rendering is mostly done on here so to get rid of all the useEffects and just have the tags and postcards grabbed before the user looks at the site so to make it quicker. Explore is still kinda slow since it has to get every postcard from memory. Not as bad as sending over network but still... Since the explore postcard thing is also used in the postcard page, I just made it it's own component and duplicated it in the postcard index. Made little tweak to the navBar.